lib: move encodeStr function to internal for reusable
